+++ mobile/android/android-components/docs/changelog.md
@@ -44,6 +44,11 @@ permalink: /changelog/
 * **feature-tabs**
   * ⚠️ **This is a breaking change**: `TabsFeature` now supports providing custom use case implementations. Therefore, an instance of `SelectTabUseCase` and `RemoveTabUseCase` have to be provided.
 
+* **support-ktx**
+  * Added `Uri.sameHostAs` to check if two Uris have the same host (both http/https, both same domain).
+  * Added `Uri.sameOriginAs` to check if two Uris have the same origin (same host, same port).
+  * Added `Uri.isInScope` to check if a Uri is within one of the given scopes.
+
 # 44.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v43.0.0...v44.0.0)